Fix troePort type mismatch: unsigned short (2 bytes) vs PaInt (4 bytes)
ASan detected global-buffer-overflow because troePort was declared as
unsigned short but the parseArgs table used PaInt, causing a 4-byte
write to a 2-byte variable. Changed to int in both declaration and
extern.

+++ src/app/orionld/orionld.cpp
@@ -231,7 +231,7 @@ bool            pernot;
 bool            disableFileLog;
 bool            lmtmp;
 char            troeHost[256];
-unsigned short  troePort;
+int             troePort;
 char            troeUser[256];
 char            troePwd[256];
 char            troeSslMode[64];
